bug fix generate code response

+++ domain/src/main/java/com/pmb/domain/model/LoginResponse.kt
@@ -26,7 +26,7 @@ data class LoginResponse(
     val lockDate: Int,
     val lockTime: Int,
     val name: String?,
-    val nationalCode: Int,
+    val nationalCode: Long,
     val passportNumber: String?,
     val passwordX: String?,
     val postalCode: String?,
